Separate proper applications having a main method from the other application classes.
