+++ src/main/java/ro/sellfluence/apphelper/GetOverview.java
@@ -1,4 +1,4 @@
-package ro.sellfluence.app;
+package ro.sellfluence.apphelper;
 
 import ro.sellfluence.googleapi.DriveAPI;
 import ro.sellfluence.googleapi.SheetsAPI;
@@ -54,4 +54,3 @@ public class GetOverview {
                 .toList();
     }
 }
-
