+++ src/main/java/ro/sellfluence/apphelper/GetStatsForAllSheets.java
@@ -1,4 +1,4 @@
-package ro.sellfluence.app;
+package ro.sellfluence.apphelper;
 
 import ro.sellfluence.googleapi.SheetsAPI;
